+++ jcache/infinispan-externalizer/src/main/java/org/javaee8/jcache/infinispan/externalize/KeyValueService.java
@@ -1,4 +1,4 @@
-package org.javaee8.jcache.configuration;
+package org.javaee8.jcache.infinispan.externalize;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
