+++ jcache/infinispan-externalizer/src/main/java/org/javaee8/jcache/infinispan/externalize/MyValue.java
@@ -1,4 +1,4 @@
-package org.javaee8.jcache.configuration;
+package org.javaee8.jcache.infinispan.externalize;
 
 import java.io.Serializable;
 import java.util.Objects;
